+++ src/js/fable-utils/RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 1.0.4
+
+* Add `validateFableOptions` helper
+
 ### 1.0.2
 
 * Add `resolveBabelOptions` helper